requirements.txt
streamlit==1.38.0
numpy==1.23.5
gtts==2.3.0
pandas==2.1.0
altair==4.2.0
scikit-learn==1.2.2
nltk==3.8.1
fuzzywuzzy==0.18.0
Pillow==9.5.0
keras==3.2.0
tensorflow==2.17.0
streamlit-option-menu
python-Levenshtein
pytesseract
opencv-python
